requirements.txt
numpy==1.26.4
pandas==2.2.2
torch==2.2.2  --index-url https://download.pytorch.org/whl/cu121
soupsieve==2.6
beautifulsoup4==4.12.3
chromadb==0.5.13
PyMuPDF==1.24.4
python-docx==1.1.2
openpyxl==3.1.5
readability-lxml==0.8.1
html-text==0.6.2
Markdown==3.7
FlagEmbedding==1.2.10
playwright==1.44.0
PyYAML==6.0.1
rwkv==0.8.26
streamlit==1.39.0   # delete
psutil==6.1.0
fastapi==0.115.4